chore: move i18n check script
